Working out build system
I think I may need to transistion to having two pyproject.tomls for
spiceypy and cyice which would require some changes to the layout and
possibly duplication of get_cspice.py (or make that a third project that
all others depend on).
