+++ setup.cfg
@@ -57,7 +57,7 @@ tests = numpy>=1.17.0; pytest>=7.0.0; pandas>=0.24.0
 dev = numpy>=1.17.0; pytest>=7.0.0; pandas>=0.24.0; coverage>=5.1.0; codecov>=2.1.0; twine>=3.3.0; wheel; build; black;
 
 [options.package_data]
-* = get_spice.py, LICENSE, README.rst, *.so, *.dll, *.dylib
+* = get_spice.py, build_spiceypy.py, LICENSE, README.rst, *.so, *.dll, *.dylib
 spiceypy = utils/*.so, utils/*.dylib, utils/*.dll
 
 [coverage:run]
