+++ setup.py
@@ -23,120 +23,6 @@ SOFTWARE.
 """
 __author__ = "AndrewAnnex"
 
-from setuptools import setup, Command, find_packages
-from setuptools.command.install import install
-from setuptools.command.build_py import build_py
-from setuptools.dist import Distribution
-import os
-import sys
-from pathlib import Path
+from setuptools import setup
 
-passnumber = 0
-
-
-def try_get_spice():
-    global passnumber
-    print("try_get_spice", passnumber)
-    if passnumber > 0:
-        print("already built libcspice")
-        return
-    try:
-        thisfile = Path(__file__).resolve(strict=False)
-        curdir = thisfile.parent
-        sys.path.append(str(curdir))
-        from get_spice import main
-
-        main()
-        passnumber += 1
-    except Exception as e:
-        print("Caught file not found")
-        print(e)
-        pass
-    except ModuleNotFoundError as mnfe:
-        print("Could not import try_get_spice")
-        pass
-    return
-
-
-class SpiceyPyBinaryDistribution(Distribution):
-    def is_pure(self):
-        return False
-
-    def root_is_pure(self):
-        return False
-
-
-class InstallSpiceyPy(install):
-    """Class that extends the install command and encapsulates the
-    process for installing the required CSPICE distribution at the
-    right place.
-    """
-
-    def finalize_options(self):
-        super().finalize_options()
-        self.install_lib = self.install_platlib
-
-    def run(self):
-        try_get_spice()
-        return super().run()
-
-
-class BuildPyCommand(build_py):
-    """Custom build command to ensure cspice is built and packaged"""
-
-    def run(self):
-        try_get_spice()
-        return super().run()
-
-
-cmdclass = {
-    "install": InstallSpiceyPy,
-    "build_py": BuildPyCommand,
-}
-
-# https://stackoverflow.com/questions/45150304/how-to-force-a-python-wheel-to-be-platform-specific-when-building-it
-# http://lepture.com/en/2014/python-on-a-hard-wheel
-try:
-    from wheel.bdist_wheel import bdist_wheel as _bdist_wheel
-
-    class generic_bdist_wheel(_bdist_wheel):
-        """
-        override for bdist_wheel
-        """
-
-        root_is_pure = False
-
-        def finalize_options(self) -> None:
-            self.root_is_pure = False
-            super().finalize_options()
-            self.root_is_pure = False
-
-        def get_tag(self) -> (str, str, str):
-            self.root_is_pure = False
-            python, abi, plat = super().get_tag()
-            self.root_is_pure = False
-            return "py3", "none", plat
-
-    # add our override to the cmdclass dict so we can inject this behavior
-    cmdclass["bdist_wheel"] = generic_bdist_wheel
-
-except ImportError:
-    # we don't have wheel installed so there is nothing to change
-    pass
-
-
-readme = open("README.rst", "r")
-readmetext = readme.read()
-readme.close()
-
-# todo: https://setuptools.pypa.io/en/latest/userguide/extension.html,
-# https://setuptools.pypa.io/en/latest/deprecated/distutils/extending.html?highlight=cmdclass#integrating-new-commands
-
-setup(
-    distclass=SpiceyPyBinaryDistribution,
-    packages=find_packages("src"),
-    package_dir={"": "src"},
-    package_data={"spiceypy": ["utils/*.so", "utils/*.dylib", "utils/*.dll"]},
-    include_package_data=True,
-    cmdclass=cmdclass,
-)
+setup()
